+++ app/src/main/java/org/linphone/ui/main/chat/model/FileModel.kt
@@ -45,6 +45,7 @@ class FileModel
     val fileCreationTimestamp: Long,
     val isEncrypted: Boolean,
     val originalPath: String,
+    val isFromEphemeralMessage: Boolean,
     val isWaitingToBeDownloaded: Boolean = false,
     val flexboxLayoutWrapBefore: Boolean = false,
     private val onClicked: ((model: FileModel) -> Unit)? = null